requirements.txt
APScheduler==3.11.1
Authlib==1.6.5
blinker==1.9.0
cachelib==0.13.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.2.0
cryptography==46.0.3
cycler==0.12.1
Flask==3.1.2
Flask-OAuthlib==0.9.6
fonttools==4.47.0
git-filter-repo==2.47.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.5
MarkupSafe==3.0.3
matplotlib==3.8.2
mysql-connector-python==8.2.0
numpy==1.26.2
oauthlib==2.1.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.21.12
pycparser==2.23
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.2.1
pytz==2023.3.post1
requests==2.32.5
requests-oauthlib==1.1.0
secure-smtplib==0.1.1
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
